requirements.txt
Pillow>=9.0.0
pyinstaller>=5.0.0
pywin32>=227; sys_platform == "win32"
winshell>=0.6; sys_platform == "win32"
